% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibleverse-package.R
\docType{package}
\name{biblegatewayr-package}
\alias{biblegatewayr}
\alias{biblegatewayr-package}
\title{biblegatewayr: Retrieve Bible Verses from Bible Gateway}
\description{
| Provides functions to interact with the BibleGateway web page. It includes 'get_bible_verse' to retrieve specific verses by canonical reference (e.g., "John 3:16") and 'find_bible_verse' to search for verses containing specific text. Both functions support multiple Bible versions and return formatted output, making it easy to incorporate Bible verses into 'markdown documents'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/angelfelizr/biblegatewayr}
  \item \url{https://angelfelizr.github.io/biblegatewayr/}
  \item Report bugs at \url{https://github.com/angelfelizr/biblegatewayr/issues}
}

}
\author{
\strong{Maintainer}: Angel Feliz \email{angel.esteban.feliz@gmail.com} (\href{https://orcid.org/0009-0007-2910-1808}{ORCID}) [copyright holder]

}
\keyword{internal}
