% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bible_verse.R
\name{get_bible_verse}
\alias{get_bible_verse}
\title{Get a Specific Bible Verse from BibleGateway}
\usage{
get_bible_verse(search, version = "RVR1960")
}
\arguments{
\item{search}{A character string representing the Bible verse reference to search for (e.g., "John 3:16", "Psalm 23").
Spaces will be converted to '+' for the URL.}

\item{version}{A character string specifying the Bible version to use (e.g., "RVR1960", "NIV", "ESV").
Defaults to "RVR1960".}
}
\value{
Invisibly returns a character string with the formatted Bible verse, or NULL if no verse is found.
The formatted verse is also printed to the console.
}
\description{
This function retrieves a specific Bible verse using the BibleGateway web site
based on a verse reference query. It returns the formatted verse text.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get a specific verse (printing to console)
get_bible_verse("Juan 3:16")

# Get a verse with a different version and capture the output
result <- get_bible_verse("Salmos 23")
cat(result)

# Example of no results.
get_bible_verse("this is a non existing book 1:1")
\dontshow{\}) # examplesIf}
}
