% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchBible-package.R
\docType{package}
\name{SearchBible-package}
\alias{SearchBible}
\alias{SearchBible-package}
\title{SearchBible: Tools for Accessing Bible Verses from BibleGateway}
\description{
| This R package provides functions to retrieve Bible verses from BibleGateway. The 'find_bible_verse' function allows you to search for Bible verses containing specific text, while the 'get_bible_verse' function enables retrieval of verses based on their canonical reference (e.g., "John 3:16"). Both functions support multiple Bible versions and provide formatted output.
}
\author{
\strong{Maintainer}: Angel Feliz \email{angel.esteban.feliz@gmail.com}

}
\keyword{internal}
